Split document service unit tests into separate files
